+++ modules/base/GSDynamicStyle.mjs
@@ -21,42 +21,60 @@ export class GSDynamicStyle extends CSSStyleSheet {
         this.id = id || GSID.id;
     }
 
+    /**
+     * Convert id to CSS class name .[name] or [name] if raw
+     * @param {String} id 
+     * @param {boolean} raw If true, id is used as is, otherwise prefixed with '.'
+     * @returns 
+     */
+    #toName(id = '', raw = false) {
+        return raw ? id : `.${id}`;
+    }
+
 	/**
 	 * Remove dynamic element rule
 	 * @param {String} id  
+     * @param {boolean} raw If true, id is used as is, otherwise prefixed with '.'
 	 */
-	deleteRule(id = '') {
+	deleteRule(id = '', raw = false) {
         const me = this;
+        const name = me.#toName(id, raw);
 		Array.from(me.cssRules)
-			.map((v, i) => v.selectorText === `.${id}` ? i : -1)
+			.map((v, i) => v.selectorText === name ? i : -1)
 			.filter(v => v > -1)
-			.forEach(v => me.deleteRule(v));
+			.forEach(v => super.deleteRule(v));
 	}    
 
     /**
      * Get individual rule from dynamic styles cache
      * @param {String} id CSS class name
+     * @param {boolean} raw If true, id is used as is, otherwise prefixed with '.'
      * @returns {CSSRule} 
      */
-    getRule(id = '') {
-        return Array.from(this.cssRules).filter(v => v.selectorText === `.${id}`).pop();
+    getRule(id = '', raw = false) {
+        const me = this;
+        const name = me.#toName(id, raw);
+        return Array.from(me.cssRules).filter(v => v.selectorText === name).pop();
     }
 
     /**
      * Set dynamic element rule. Accepts string or json object representation
      * @param {String} id Unique css class name used
      * @param {string|object} style CSS style properties to apply
+     * @param {boolean} sync If true, applys rules immediately, otherwise on next animation frame
+     * @param {boolean} raw If true, id is used as is, otherwise prefixed with '.'
      */
-    setRule(id, style = '', sync = false) {
+    setRule(id, style = '', sync = false, raw = false) {
 
         const me = this;
 
         if (!id) return;
 
-		let rule = me.getRule(id);
+		let rule = me.getRule(id, raw);
 		if (!rule) {
-			me.insertRule(`.${id} {}`);
-			return me.setRule(id, style, sync);
+            const name = me.#toName(id, raw);
+			me.insertRule(`${name} {}`);
+			return me.setRule(id, style, sync, raw);
 		}        
 
         if (!style) return rule;
@@ -108,47 +126,51 @@ export class GSDynamicStyle extends CSSStyleSheet {
      * @returns {CSSStyleSheet} An instance of dynamic stylesheet containing dynamic rules
      */
     static dynamicStyleSheetByID(root, id) {
-		return Array.from(root?.adoptedStyleSheets).filter(el => el.id === id).pop();
+		return root?.adoptedStyleSheets ?  Array.from(root?.adoptedStyleSheets).filter(el => el.id === id).pop() : null;
     }
 
 	/**
 	 * Find dynamic style & rule for element
 	 * @param {HTMLElement} source A html element with dynamic style attached  (with data-gs-class attribute set)
+     * @param {boolean} raw If true, id is used as is, otherwise prefixed with '.'
 	 * @returns {CSSStyleRule} Instance of dynamic css class 
 	 */
-	static dynamicRule(source) {
+	static dynamicRule(source, raw = false) {
 		const style = GSDynamicStyle.dynamicStyleSheet(source);
-		return style?.getRule(source.dataset.gsClass);		
+		return style?.getRule(source.dataset.gsClass, raw);		
 	}
 
     /**
      * Fnd custom dynamic rule
      * @param {Document} root Document or shadowDom
      * @param {String} id Dynamic CSS name
-    * @returns {CSSStyleRule} Instance of dynamic css class 
+     * @param {boolean} raw If true, id is used as is, otherwise prefixed with '.'
+     * @returns {CSSStyleRule} Instance of dynamic css class 
      */    
-    static dynamicRuleByID(root, id) {
+    static dynamicRuleByID(root, id, raw = false) {
         const style = GSDynamicStyle.dynamicStyleSheetByID(root, 'dynamic');
-        return style?.getRule(id);
+        return style?.getRule(id, raw);
     }
 
     /**
      * Find element dynamic style
      * @param {HTMLElement} source 
+     * @param {boolean} raw If true, id is used as is, otherwise prefixed with '.'
      * @returns {CSSStyle} Element dynamic style or default style if dynamic not defined
      */
-	static dynamicStyle(source) {
-		const rule = GSDynamicStyle.dynamicRule(source);
+	static dynamicStyle(source, raw = false) {
+		const rule = GSDynamicStyle.dynamicRule(source, raw);
 		return rule?.style; // || source.style;
 	}
 
     /**
      * Find element dynamic style
      * @param {Document} source Document or shadowDOM
+     * @param {boolean} raw If true, id is used as is, otherwise prefixed with '.'
      * @returns {CSSStyle} Element dynamic style or default style if dynamic not defined
      */
-	static dynamicStyleByID(root, id) {
-		const rule = GSDynamicStyle.dynamicRuleByID(root, id);
+	static dynamicStyleByID(root, id, raw = false) {
+		const rule = GSDynamicStyle.dynamicRuleByID(root, id, raw);
 		return rule?.style; // || source.style;
 	}
 
@@ -157,11 +179,12 @@ export class GSDynamicStyle extends CSSStyleSheet {
      * Will search for a StyleSheet within element shadow addoptedStyleSheet by element data-gs-class property.
      * @param {HTMLElement} source Element with dynamic style attached (with data-gs-class attribute set)
      * @param {Object} opt A JSON object of css arguments to apply 
+     * @param {boolean} raw If true, id is used as is, otherwise prefixed with '.'
      * @returns {CSSStyle} A copy of modifid CSSStyle object 
      */
-	static applyDynamicStyle(source, opt) {
-        requestAnimationFrame(() => {
-            const style = GSDynamicStyle.dynamicStyle(source);
+	static applyDynamicStyle(source, opt, raw = false) {
+        queueMicrotask(() => {
+            const style = GSDynamicStyle.dynamicStyle(source, raw);
             return Object.assign(style, opt);
         });
 	}
@@ -171,11 +194,12 @@ export class GSDynamicStyle extends CSSStyleSheet {
      * @param {Document || ShadowRoot} root Element with dynamic style attached
      * @param {String} id Dynamic CSS style name
      * @param {Object} opt A JSON object of css arguments to apply 
+     * @param {boolean} raw If true, id is used as is, otherwise prefixed with '.'
      * @returns {CSSStyle} A copy of modifid CSSStyle object 
      */
-	static applyDynamicStyleByID(root, id, opt) {
-        requestAnimationFrame(() => {
-            const style = GSDynamicStyle.dynamicStyleByID(root, id);
+	static applyDynamicStyleByID(root, id, opt, raw = false) {
+        queueMicrotask(() => {
+            const style = GSDynamicStyle.dynamicStyleByID(root, id, raw);
             return Object.assign(style, opt);
         });
 	}